11-misc/requirements.txt
flask
flask_smorest
flask-sqlalchemy
pymysql
marshmallow
flask-jwt-extended
passlib
